refactor: Replace theme.ts for themes/index.ts
