+++ packages/web/src/lib/fonts.ts
@@ -1,11 +1,12 @@
 export const fonts = {
-    "JetBrains": "JetBrains Mono",
-    "Monaco": "Monaco",
-    "Courier": "Courier",
-    "MonoCraft": "Monocraft Nerd Font",
-    "BigBlue Terminal": "BigBlue TerminalPlus",
-    "JGS": "jgs_Font",
-    "Steps Mono": "Steps Mono",
-    "IBM Plex Mono": "IBM Plex Mono",
-    "Inconsolata": "Inconsolata",
-}
+  JetBrains: "JetBrains Mono",
+  Monaco: "Monaco",
+  Courier: "Courier",
+  MonoCraft: "Monocraft Nerd Font",
+  "BigBlue Terminal": "BigBlue TerminalPlus",
+  JGS: "jgs_Font",
+  "Steps Mono": "Steps Mono",
+  "IBM Plex Mono": "IBM Plex Mono",
+  Inconsolata: "Inconsolata",
+};
+export default fonts;
