--- packages/web/src/lib/theme.ts
@@ -1,14 +1,0 @@
-import { codeMirrorThemes } from "./themes"
-
-type Themes = {
-  [key: string]: any
-}
-
-export const themes: Themes = {
-  "dracula": codeMirrorThemes.dracula,
-  "monokai": codeMirrorThemes.monokai,
-  "gruvbox": codeMirrorThemes.gruvbox,
-  "ayuDark": codeMirrorThemes.ayuDark,
-  "tokyo Night": codeMirrorThemes.tokyonight,
-  "nord": codeMirrorThemes.nord,
-}
