+++ packages/web/src/lib/themes/index.ts
@@ -2,14 +2,20 @@ import { ayuDark } from "./ayu-dark";
 import { dracula } from "./dracula";
 import { gruvboxDark } from "./gruvbox-dark";
 import { monokai } from "./monokai-dimmed";
-import { tokyoNight } from "./tokyo-night";
 import { nord } from "./nord";
+import { tokyoNight } from "./tokyo-night";
 
-export const codeMirrorThemes = {
-    "dracula": dracula,
-    "monokai": monokai,
-    "gruvbox": gruvboxDark,
-    "ayuDark": ayuDark,
-    "tokyonight": tokyoNight,
-    "nord": nord
+type Themes = {
+  [key: string]: any;
 };
+
+export const themes: Themes = {
+  ayuDark,
+  dracula,
+  gruvboxDark,
+  monokai,
+  nord,
+  tokyoNight,
+};
+
+export default themes;
